migrations: Enforce that they must be backwards compatible.
Since migrations may happen while traffic is still being served by the
previous version, it is important that they be backwards-compatible;
that is, no column or table drops or renames unless production has
already stopped relying on them.

Add a linter which validates this property.  Since there have
_already_ been non-backwards-compatible changes (which were fine, it's
only going forward that matters!), we pass it a "high water mark" of
a recent commit on main; only migrations after that point are checked.

Because this relies on commit history in CI, it requires a non-shallow
clone in the github checkout action.  Particularly for small
repositories, shallow clones are significantly more CPU intensive, for
only tiny gain; switching to a deep clone is not a notable cost.