config: Unescape colons in file paths.
In Google Cloud, connections to Cloud SQL are made available in the
filesystem, at paths like `/cloudsql/projectid:region:dbinstance`.
Normally, file paths are passed in via DATABASE_URL in the form
`postgres://user:password@%2Fvar%2Flib%2Fpostgresql/dbname`.  In this
format, a file path with colons in it has the colon parsed as a
delimiter for a _port_, not as part of the file path.

This same problem, but with slashes, is solved by URI-escaping them as
`%2f`.  Unfortunately, `dj-database-url` does not support
un-URI-escaping the colons (jacobian/dj-database-url#113), so the
unescaping must be done by hand in the configuration file.