sentry: Create release and deploy markers.
See https://docs.sentry.io/product/releases/ for details.

We use an unusual construction of an explicit `/bin/sh` entrypoint for these
steps, because `${VARS}` are substituted _before_ the shell, and
secretEnv are actually substituted as environment variables -- which
means the shell must be used to give a chance to interpolate them.