sentry: Associate releases with commits.
The default clone is a shallow one[1], which makes Sentry unable to
automatically determine which commits are new in a given release.
Despite being able to clone the repository already, via OAuth scopes,
we must provide our own deploy key[2] in order to deepen the shallow
clone to get full history.

[1] https://cloud.google.com/build/docs/automating-builds/create-manage-triggers#including_the_repository_history_in_a_build
[2] https://cloud.google.com/build/docs/access-private-github-repos