import_counties: Use a single insert, for efficiency.
A large number of separate insert statements is inefficient.  It is
made more so by running `update_or_create` to check each row, when
this import process is the only one which creates county objects.

Wrap the creation in a transaction, and bulk-create all counties with
a single insert.  We take an update lock on any existing counties
beforehand, ensuring that any two invocations of this importer run
serially; other (currently hypothetical) codepaths that create
counties may cause the entire transaction to abort if they insert
rows, but this will at least lead to consistent state.