core: Lengthen the default 50-char limit for Reporter external_id.
The `external_id` contains a bunch of baggage that can make it quite
long:

    auth0:samlp|SomeOrg|first.middle.othermiddle.lastname@longdomain.com

This can make it impossible for some users to log in, based on having
too long an email address.

Lengthen the slug size to 400 characters.  The maximum size of a
`varchar` does not affect its storage in PostgreSQL, and email
addresses, per RFC, can be up to 320 characters long.

Because this lengthens the field, it is backwards-compatible; it
cannot cause uniqueness conflicts.