JWT now verified without an extra HTTP hit, refs #131
It's actually a bit more complicated than that.

We now decode the JWT from the SPA in the same way as the
JWT from the Django admin sign-in flow.

But... the JWT from the SPA does not include the name and email.

So the first time we see a JWT for a new user we fetch their
name and email from the userinfo endpoint to create their
record, but on subsequent JWT decodes we do not need to do that.