submitReport: Capture if the report was originally flagged for review.
Because the Journeyman role is randomly chosen for review 15% of the
time, and we have no way to log changes to the "pending review" flag,
it is not possible to report on if a particular report was
_originally_ flagged for review, as that is not predictable from the
original `api_log`, nor accurate from the current `report`.

Add an immutable-in-the-UI column for if the report was originally
chosen for review.  We may have better solutions in the
future (e.g. using revision objects) but this is a quick and dirty
stopgap.

A stopgap fix for #274.