report: Factor out a full_appointment_details for MyTurn / county types.
The behaviour of looking up the county website, or returning the
MyTurn website, is going to be common across many exporters.  Instead
of placing this logic just in the API exporter, move it onto the
model.

To prevent additional fetches, the method supports being passed in the
record's location.  This is validated to be the right location, which
does mean that the exporter query gets one column wider, which is a
negligible cost.