model: Update denormalizations, even during bulk update.
The only code that updated denormalizations after unsetting
is_pending_review lay in the Report admin view; as such, bulk updates
which added Approved and unset the flag failed to update the
denormalized reports on locations.

Rather than continue to play whack-a-mole with this, add a signal
receiver for the Approved note, which bubbles up to unset the
is_pending_review flag; this, in turn, will trigger a denormalization
update.

This is slightly less performant than putting the denormalization
update into the bulk update, but guarantees that we catch other
related failures in the future.