auth: Check user permissions.
Except when using a VIAL auth token (which is an id token, and thus
does not have permissions in it, but is only used for debugging)
validate that the user has the right permissions on the API.  For now,
this is always the "caller" API.

By checking this on every request, resolves #393; the user would have
been forced to log out and back in, thus capturing their new roles.

Side-steps #211, which is about adding scopes to the _internal_ API
keys abstraction, which feels like a step in the wrong direction; we
already have a complicated permissions model via Auth0, and adding
another one atop of it will not help.  Additionally, browser clients
are unable to make use of `ApiKey`s, since we have no way to give them
out.

Fixes #393.
Fixes #187.