auth: Update user metadata on every call stats request.
Updating the user (name and email) metadata requires making a request
to Auth0, which we only used to do on user creation.  Because this
never got updated, it meant that various workarounds (#423) were put
in place to allow users to be superficially "renamed" in VIAL to match
Auth0.

Pull updated user information from Auth0 on the stats request, which
is done at the start of every calling run.  This does not _guarantee_
up-to-date information in VIAL, as Auth0 may have been updated since
the last login in VIAL, but makes it at least eventually-consistent.

Fixes #229.