call_request: Many fixes; uncompleted locations are unique in the queue.
This adds a unique partial index on location, for non-completed
call_requests, to ensure that a location cannot be in the queue more
than once.  It adds a "validity" filter to locations, based on having
a phone number, being not marked do-not-call, being not soft-deleted,
and being not marked for online research, and consistently enforces
that no such locations can enter into the call request list, nor
remain there if they change into being so[1].
Fixes #325.
Fixes #295.

We switch to using class methods on the CallRequest class to insert,
claim, report results of, and backfill, and use those consistently in
all callpaths.

This fixes a bug wherein skipped call requests would be placed at the
_top_ of the list for their priority group, rather than the bottom.
This updates the behaviour in #332 to be as expected.

Since locations can only occur once in the uncompleted call request
list, we always attach a report to any uncompleted call report for its
location, if there is one.  This resolves the issue where a skip would
file a pending call request, and the reporter would edit it,
submitting a report -- but the unvested call request would still be
pending.  We now associate the "edit" with the unvested call request
and resolve it.
Fixes #357.

Enables #502 -- it is now a set which tags allow rescheduling.

[1] One notable exception exists for this -- usage of `.update()`
bypasses triggers, so can cause a non-valid location to remain in the
list.  Bulk updates of locations are quite rare, so we'll observe and
address this more (in get_call_request) if it is an issue.