admin: Large hammer to disable revert-to-revision.
Changes in reversion are tracked by "version" objects tied to a
record, which are grouped into "revision" objects which may span
multiple objects. When importing, we bulk-create locations in batches
of 20, which results in a revision with a version for each location,
which is reasonable.

Reverting a location re-applies all changes in the _revision_, not
just the _version_.  This means that pressing "revert to revision" on
a location back to an import revision will _silently_ change 19 other
locations at the same time back to their import-time values, with
possibly disastrous consequences for database integrity.

Entirely prevent the POST request that would do this.  Subtler
solutions exist, but this large hammer is good enough for now, and
leaves the version history browsable, which provides value still.